--- packages/react-img-viewer/.github/workflows/size.yml
@@ -1,12 +1,0 @@
-name: size
-on: [pull_request]
-jobs:
-  size:
-    runs-on: ubuntu-latest
-    env:
-      CI_JOB_NUMBER: 1
-    steps:
-      - uses: actions/checkout@v1
-      - uses: andresz1/size-limit-action@v1
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
