--- packages/react-img-viewer/.gitignore
@@ -1,8 +1,0 @@
-*.log
-.DS_Store
-node_modules
-.cache
-dist
-src/stories
-
-*storybook.log
