--- packages/react-img-viewer/example/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-.cache
-dist
