--- packages/react-img-viewer/package.json
@@ -1,74 +1,0 @@
-{
-  "version": "0.1.0",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test --passWithNoTests",
-    "lint": "tsdx lint",
-    "prepare": "tsdx build",
-    "size": "size-limit",
-    "analyze": "size-limit --why",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
-  },
-  "peerDependencies": {
-    "react": ">=16"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "tsdx lint"
-    }
-  },
-  "prettier": {
-    "printWidth": 80,
-    "semi": true,
-    "singleQuote": true,
-    "trailingComma": "es5"
-  },
-  "name": "react-img-viewer",
-  "author": "manav",
-  "module": "dist/react-img-viewer.esm.js",
-  "size-limit": [
-    {
-      "path": "dist/react-img-viewer.cjs.production.min.js",
-      "limit": "10 KB"
-    },
-    {
-      "path": "dist/react-img-viewer.esm.js",
-      "limit": "10 KB"
-    }
-  ],
-  "devDependencies": {
-    "@size-limit/preset-small-lib": "^11.1.4",
-    "@storybook/addon-essentials": "^8.1.10",
-    "@storybook/addon-interactions": "^8.1.10",
-    "@storybook/addon-links": "^8.1.10",
-    "@storybook/addon-onboarding": "^8.1.10",
-    "@storybook/addon-webpack5-compiler-swc": "1.0.3",
-    "@storybook/blocks": "^8.1.10",
-    "@storybook/react": "^8.1.10",
-    "@storybook/react-webpack5": "^8.1.10",
-    "@storybook/test": "^8.1.10",
-    "@types/iv-viewer": "^2.0.3",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "husky": "^9.0.11",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "size-limit": "^11.1.4",
-    "storybook": "^8.1.10",
-    "tsdx": "^0.14.1",
-    "tslib": "^2.6.3",
-    "typescript": "^5.4.5"
-  }
-}
