+++ packages/react-iv-viewer/.gitignore
@@ -1,0 +1,7 @@
+*.log
+.DS_Store
+node_modules
+.cache
+dist
+
+*storybook.log
