+++ packages/react-iv-viewer/README.md
@@ -1,0 +1,44 @@
+# Getting Started
+
+`react-iv-viewer` is a React-based library for viewing images with advanced features like zooming, high-resolution image support, and full-screen viewing. It provides an easy way to integrate image viewing functionality into your React applications.
+
+### Features
+
+- Smooth dragging and panning of images
+- Support for touch devices
+- Double tap to zoom in/zoom out
+- Pinch in/out to control zoom
+- Zooming with mouse wheel or buttons
+- High-resolution image support
+- Full-screen image viewing
+- Customizable viewer dimensions
+- Event handlers for zoom and load events
+- TypeScript support
+
+### Demos
+
+See the many DEMO sections in the documentation.
+
+### Install
+
+using `npm`
+```bash
+npm install react-iv-viewer
+```
+
+using `yarn`
+```bash
+yarn add react-iv-viewer
+```
+
+#### ES6
+
+Image Viewer
+```jsx
+import { ImageViewer } from 'react-iv-viewer';
+```
+
+Fullscreen Viewer
+```jsx
+import { FullScreenImageViewer } from 'react-iv-viewer';
+```
