+++ packages/react-iv-viewer/global.d.ts
@@ -1,0 +1,27 @@
+// global.d.ts
+
+declare module '*.png' {
+    const value: string;
+    export default value;
+  }
+  
+  declare module '*.jpg' {
+    const value: string;
+    export default value;
+  }
+  
+  declare module '*.jpeg' {
+    const value: string;
+    export default value;
+  }
+  
+  declare module '*.gif' {
+    const value: string;
+    export default value;
+  }
+  
+  declare module '*.svg' {
+    const value: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
+    export default value;
+  }
+
